Move `QueueStoreRepository` tests into `providertest`.
