+++ persistence/internal/providertest/suite.go
@@ -5,7 +5,6 @@ import (
 	"time"
 
 	"github.com/dogmatiq/infix/persistence/internal/providertest/common"
-	"github.com/dogmatiq/infix/persistence/internal/providertest/queuestore"
 	marshalkitfixtures "github.com/dogmatiq/marshalkit/fixtures"
 	"github.com/onsi/ginkgo"
 )
@@ -63,8 +62,8 @@ func Declare(
 		declareEventOperationTests(&tc)
 		declareEventRepositoryTests(&tc)
 
-		queuestore.DeclareRepositoryTests(&tc)
 		declareQueueOperationTests(&tc)
+		declareQueueRepositoryTests(&tc)
 
 		declareOffsetOperationTests(&tc)
 		declareOffsetRepositoryTests(&tc)
